feat: New functions for resending email codes
Deleted pnpm because it was breaking docker build

Added function for resendingEmailCodes, user can in duration of 300
seconds send 2 requests for new emailCode

Added simple toast ui for notifying user of exceeding allowed amount of
requests for resending email code

+++ src/lib/emailCode.ts
@@ -1,12 +1,25 @@
 import { TimeSpan, createDate, isWithinExpirationDate } from "oslo";
 import { generateRandomString, alphabet } from "oslo/crypto";
 import { prisma } from "./db";
-import type { User } from "lucia";
+import { emailCodeRateLimiterMiddleware } from "./RateLimiter";
+
+interface EmailVerificationCodeReturnType {
+  data: string | null;
+  error: string | null;
+}
 
 export async function generateEmailVerificationCode(
   userId: string,
   email: string,
-): Promise<string> {
+  ip: string
+): Promise<EmailVerificationCodeReturnType> {
+
+  const emailLimiter = await emailCodeRateLimiterMiddleware({ remoteAddress: ip });
+  console.log (emailLimiter);
+  if (!emailLimiter) {
+    return { data: null, error: "Too many requests" };
+  }
+
   await prisma.emailVerificationCode.deleteMany({
     where: {
       userId,
@@ -23,5 +36,5 @@ export async function generateEmailVerificationCode(
     },
   });
 
-  return code;
+  return { data: code, error: null };
 }
